# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.createTaskRequest:
    properties:
      git_url:
        type: string
    type: object
...
  model.GenTask:
    properties:
      created_at:
        type: string
      error:
        type: string
      git_url:
        type: string
      id:
        type: integer
      message:
        type: string
      progress:
        type: integer
      queued_at:
        description: |-
          QueuedAt 消息成功投递到 MQ 的时刻；NULL 表示尚未投递成功（孤儿任务，
          由启动扫描补投），是「落库」与「投递」两步之间崩溃的补偿依据。
        type: string
      repo_analysis_id:
        type: integer
      status:
        type: string
      step:
        type: string
      updated_at:
        type: string
    type: object
  resp.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.PageData:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  service.TaskDetail:
    properties:
      created_at:
        type: string
      error:
        type: string
      git_url:
        type: string
      id:
        type: integer
      logs:
        items:
          type: string
        type: array
      message:
        type: string
      progress:
        type: integer
      queued_at:
        description: |-
          QueuedAt 消息成功投递到 MQ 的时刻；NULL 表示尚未投递成功（孤儿任务，
          由启动扫描补投），是「落库」与「投递」两步之间崩溃的补偿依据。
        type: string
      repo_analysis_id:
        type: integer
      status:
        type: string
      step:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    AI 博客生成平台后端：管理端（任务/文章/RAG 向量索引）+ 前台门户（文章浏览/RAG 问答）。
    统一响应结构 resp.Envelope{code, message, data}，code=0 表示成功；业务错误 code 为 HTTP 状态码。
    /metrics 为 Prometheus 抓取端点，不在 /api/v1 之下，本文档未收录。
  title: AI 博客平台 API
  version: "1.0"
paths:
...
  /tasks:
    get:
      parameters:
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认 10，上限 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/service.PageData'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.GenTask'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 任务分页列表
      tags:
      - 任务
    post:
      consumes:
      - application/json
      description: 克隆公开 Git 仓库 → 静态分析 → LLM 生成 4 篇中文文章（含 SVG 配图），落库为草稿。
      parameters:
      - description: Git 仓库地址
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.GenTask'
              type: object
      security:
      - BearerAuth: []
      summary: 创建生成任务
      tags:
      - 任务
  /tasks/{id}:
    delete:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - BearerAuth: []
      summary: 删除任务
      tags:
      - 任务
    get:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/service.TaskDetail'
              type: object
      security:
      - BearerAuth: []
      summary: 任务详情
      tags:
      - 任务
  /tasks/{id}/cancel:
    post:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.GenTask'
              type: object
      security:
      - BearerAuth: []
      summary: 取消任务
      tags:
      - 任务
  /tasks/{id}/retry:
    post:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/model.GenTask'
              type: object
      security:
      - BearerAuth: []
      summary: 重跑失败任务
      tags:
      - 任务
securityDefinitions:
  BearerAuth:
    description: '管理端接口需在请求头携带登录接口返回的 JWT：`Authorization: Bearer <token>`。'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"